requirements.txt
Package                   Version
------------------------- ------------
anyio                     3.5.0
argon2-cffi               21.3.0
argon2-cffi-bindings      21.2.0
asttokens                 2.0.5
async-lru                 2.0.4
attrs                     23.1.0
Babel                     2.11.0
backcall                  0.2.0
beautifulsoup4            4.12.2
bleach                    4.1.0
Bottleneck                1.3.5
Brotli                    1.0.9
cairocffi                 1.6.1
CairoSVG                  2.7.1
causal-learn              0.1.3.8
certifi                   2023.11.17
cffi                      1.16.0
charset-normalizer        2.0.4
comm                      0.1.2
contourpy                 1.2.0
cryptography              41.0.7
cssselect2                0.7.0
cycler                    0.12.1
debugpy                   1.6.7
decorator                 5.1.1
defusedxml                0.7.1
exceptiongroup            1.0.4
executing                 0.8.3
fastjsonschema            2.16.2
fonttools                 4.47.0
graphviz                  0.20.3
idna                      3.4
ipykernel                 6.25.0
ipython                   8.15.0
jedi                      0.18.1
Jinja2                    3.1.2
joblib                    1.3.2
json5                     0.9.6
jsonschema                4.19.2
jsonschema-specifications 2023.7.1
jupyter_client            8.6.0
jupyter_core              5.5.0
jupyter-events            0.8.0
jupyter-lsp               2.2.0
jupyter_server            2.10.0
jupyter_server_terminals  0.4.4
jupyterlab                4.0.8
jupyterlab-pygments       0.1.2
jupyterlab_server         2.25.1
kiwisolver                1.4.5
llvmlite                  0.39.1
MarkupSafe                2.1.3
matplotlib                3.8.2
matplotlib-inline         0.1.6
mistune                   2.0.4
mkl-fft                   1.3.8
mkl-random                1.2.4
mkl-service               2.4.0
nbclient                  0.8.0
nbconvert                 7.10.0
nbformat                  5.9.2
nest-asyncio              1.5.6
networkx                  3.2.1
notebook_shim             0.2.3
numba                     0.56.4
numexpr                   2.8.7
numpy                     1.23.5
overrides                 7.4.0
packaging                 23.1
pandas                    2.1.4
pandocfilters             1.5.0
parso                     0.8.3
patsy                     0.5.6
pexpect                   4.8.0
pickleshare               0.7.5
pillow                    10.2.0
pip                       23.3.1
platformdirs              3.10.0
prometheus-client         0.14.1
prompt-toolkit            3.0.43
psutil                    5.9.0
ptyprocess                0.7.0
pure-eval                 0.2.2
pycparser                 2.21
pydot                     2.0.0
Pygments                  2.15.1
pyOpenSSL                 23.2.0
pyparsing                 3.1.1
PySocks                   1.7.1
python-dateutil           2.8.2
python-json-logger        2.0.7
pytz                      2023.3.post1
PyYAML                    6.0.1
pyzmq                     25.1.0
referencing               0.30.2
requests                  2.31.0
rfc3339-validator         0.1.4
rfc3986-validator         0.1.1
rpds-py                   0.10.6
scikit-learn              1.4.0rc1
SciPy                     1.12.0rc1
Send2Trash                1.8.2
setuptools                68.2.2
six                       1.16.0
sniffio                   1.2.0
soupsieve                 2.5
stack-data                0.2.0
statsmodels               0.14.2
terminado                 0.17.1
threadpoolctl             3.2.0
tigramite                 5.2.2.3
tinycss2                  1.2.1
tomli                     2.0.1
tornado                   6.3.3
tqdm                      4.66.4
traitlets                 5.7.1
typing_extensions         4.7.1
tzdata                    2023.3
urllib3                   1.26.18
wcwidth                   0.2.5
webencodings              0.5.1
websocket-client          0.58.0
wheel                     0.41.2
